clean code and add first tests
TODO: Tangent Evolver tests do not pass.
